+++ service/controller/clusternamespace/resource/clusternamespaceresources/create_test.go
@@ -6,6 +6,8 @@ import (
 	"reflect"
 	"testing"
 
+	"github.com/giantswarm/rbac-operator/service/test"
+
 	"github.com/giantswarm/k8sclient/v7/pkg/k8sclienttest"
 	"github.com/giantswarm/micrologger/microloggertest"
 	security "github.com/giantswarm/organization-operator/api/v1alpha1"
@@ -31,13 +33,13 @@ func Test_EnsureCreated(t *testing.T) {
 		{
 			name: "flawless",
 			namespaces: []*corev1.Namespace{
-				newOrgNamespace("acme"),
-				newClusterNamespace("abc0", "acme"),
-				newGenericNamespace("giantswarm"),
+				test.NewOrgNamespace("acme"),
+				test.NewClusterNamespace("abc0", "acme"),
+				test.NewGenericNamespace("giantswarm"),
 			},
-			organization: newOrganization("acme"),
+			organization: test.NewOrganization("acme"),
 			roleBindings: []*rbacv1.RoleBinding{
-				newRoleBinding(
+				test.NewRoleBinding(
 					"flux-crd-controller",
 					"org-acme",
 					map[string]string{
@@ -53,7 +55,7 @@ func Test_EnsureCreated(t *testing.T) {
 						{Kind: "ServiceAccount", Name: "source-controller", Namespace: "flux-system"},
 					},
 				),
-				newRoleBinding(
+				test.NewRoleBinding(
 					"flux-namespace-reconciler",
 					"org-acme",
 					map[string]string{
@@ -65,7 +67,7 @@ func Test_EnsureCreated(t *testing.T) {
 						{Kind: "ServiceAccount", Name: "kustomize-controller", Namespace: "flux-system"},
 					},
 				),
-				newRoleBinding(
+				test.NewRoleBinding(
 					"cluster-ns-organization-acme-write",
 					"org-acme",
 					map[string]string{
@@ -76,7 +78,7 @@ func Test_EnsureCreated(t *testing.T) {
 						{Kind: "Group", Name: "customer:acme:Employees"},
 					},
 				),
-				newRoleBinding(
+				test.NewRoleBinding(
 					"cluster-ns-organization-acme-read",
 					"org-acme",
 					map[string]string{
@@ -89,7 +91,7 @@ func Test_EnsureCreated(t *testing.T) {
 				),
 			},
 			expectedRoleBindings: []*rbacv1.RoleBinding{
-				newRoleBinding(
+				test.NewRoleBinding(
 					"flux-crd-controller",
 					"abc0",
 					map[string]string{
@@ -105,7 +107,7 @@ func Test_EnsureCreated(t *testing.T) {
 						{Kind: "ServiceAccount", Name: "source-controller", Namespace: "flux-system"},
 					},
 				),
-				newRoleBinding(
+				test.NewRoleBinding(
 					"flux-namespace-reconciler",
 					"abc0",
 					map[string]string{
@@ -117,7 +119,7 @@ func Test_EnsureCreated(t *testing.T) {
 						{Kind: "ServiceAccount", Name: "kustomize-controller", Namespace: "flux-system"},
 					},
 				),
-				newRoleBinding(
+				test.NewRoleBinding(
 					"write-in-cluster-ns",
 					"abc0",
 					map[string]string{
@@ -128,7 +130,7 @@ func Test_EnsureCreated(t *testing.T) {
 						{Kind: "Group", Name: "customer:acme:Employees"},
 					},
 				),
-				newRoleBinding(
+				test.NewRoleBinding(
 					"read-in-cluster-ns",
 					"abc0",
 					map[string]string{