+++ service/test/common.go
@@ -1,4 +1,4 @@
-package clusternamespaceresources
+package test
 
 import (
 	"fmt"
@@ -10,7 +10,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
-func newClusterNamespace(name, organization string) *corev1.Namespace {
+func NewClusterNamespace(name, organization string) *corev1.Namespace {
 	return &corev1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
 			Labels: map[string]string{
@@ -22,7 +22,7 @@ func newClusterNamespace(name, organization string) *corev1.Namespace {
 	}
 }
 
-func newGenericNamespace(name string) *corev1.Namespace {
+func NewGenericNamespace(name string) *corev1.Namespace {
 	return &corev1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: name,
@@ -30,7 +30,7 @@ func newGenericNamespace(name string) *corev1.Namespace {
 	}
 }
 
-func newOrganization(name string) *security.Organization {
+func NewOrganization(name string) *security.Organization {
 	return &security.Organization{
 		ObjectMeta: metav1.ObjectMeta{
 			Name: name,
@@ -41,19 +41,19 @@ func newOrganization(name string) *security.Organization {
 	}
 }
 
-func newOrgNamespace(name string) *corev1.Namespace {
+func NewOrgNamespace(orgName string) *corev1.Namespace {
 	return &corev1.Namespace{
 		ObjectMeta: metav1.ObjectMeta{
 			Labels: map[string]string{
 				k8smetadata.ManagedBy:    "organization-operator",
-				k8smetadata.Organization: name,
+				k8smetadata.Organization: orgName,
 			},
-			Name: fmt.Sprintf("org-%s", name),
+			Name: fmt.Sprintf("org-%s", orgName),
 		},
 	}
 }
 
-func newRoleBinding(name, namespace string, roleRef map[string]string, subjects []rbacv1.Subject) *rbacv1.RoleBinding {
+func NewRoleBinding(name, namespace string, roleRef map[string]string, subjects []rbacv1.Subject) *rbacv1.RoleBinding {
 	return &rbacv1.RoleBinding{
 		TypeMeta: metav1.TypeMeta{
 			Kind:       "RoleBinding",
@@ -72,5 +72,28 @@ func newRoleBinding(name, namespace string, roleRef map[string]string, subjects 
 			Kind:     roleRef["kind"],
 			Name:     roleRef["name"],
 		},
+	}
+}
+
+func NewClusterRole(name string, rules ...rbacv1.PolicyRule) *rbacv1.ClusterRole {
+	return &rbacv1.ClusterRole{
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "ClusterRole",
+			APIVersion: "rbac.authorization.k8s.io/v1",
+		},
+		ObjectMeta: metav1.ObjectMeta{
+			Labels: map[string]string{},
+			Name:   name,
+		},
+		Rules: rules,
+	}
+}
+
+func NewPolicyRule(verbs, resources, resourceNames, apiGroups []string) *rbacv1.PolicyRule {
+	return &rbacv1.PolicyRule{
+		Verbs:         verbs,
+		Resources:     resources,
+		ResourceNames: resourceNames,
+		APIGroups:     apiGroups,
 	}
 }